изменяет функции для работы с компонентами
