+++ src/components/sort.js
@@ -1,4 +1,3 @@
-import {createElement} from "../utils";
 import AbstractComponent from "./abstract-component";
 
 const createSortTemplate = () => {
