+++ src/const.js
@@ -17,10 +17,6 @@ export const MONTH_NAMES = [
   `November`,
   `December`
 ];
-export const RENDER_POSITION = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
 export const KEY = {
   ENTER: `Enter`,
   ESC: `Escape`
