--- src/utils.js
@@ -1,22 +1,0 @@
-import {RENDER_POSITION} from './const';
-
-export const getRandomBool = () => Math.random() >= 0.5;
-export const getRandomNumber = (end, start = 0) => Math.floor(start + Math.random() * (end + 1 - start));
-export const getRandomElement = (arrayElements) => arrayElements[getRandomNumber(arrayElements.length - 1)];
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RENDER_POSITION.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RENDER_POSITION.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
