+++ src/utils/render.js
@@ -1,0 +1,37 @@
+export const RENDER_POSITION = {
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`
+};
+
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
+
+export const render = (container, component, place) => {
+  switch (place) {
+    case RENDER_POSITION.AFTERBEGIN:
+      container.prepend(component.getElement());
+      break;
+    case RENDER_POSITION.BEFOREEND:
+      container.append(component.getElement());
+      break;
+  }
+};
+
+export const replace = (newComponent, oldComponent) => {
+  const newElement = newComponent.getElement();
+  const oldElement = oldComponent.getElement();
+  const parentElement = oldComponent.getElement().parentElement;
+
+  const isElementsExist = !!(newElement && oldElement && parentElement);
+
+  if (isElementsExist && parentElement.contains(oldElement)) {
+    parentElement.replaceChild(newElement, oldElement);
+  }
+};
+
+export const remove = (component) => {
+  component.getElement().remove();
+};
